NIFIREG-58 Add current user permissions
For authorizable resources, such as tenants and buckets and bucket
items, add a permissions fields that indicates actions available to the
current user such as read, write, delete.

Adds permission details to the response of the GET /access endpoint
of the Web API that shows the available action to top level resources,
such as /buckets, /tenants, and /policies.

Add a configurable flag to tenants indicating if they are configurable
or read-only based on the Authorizer & UserGroupProvider that is
managing them.