NIFIREG-226: Bugfix overriding flow provider in Docker container via environment variables
The script which maps env to flow provider properties is currently not overwriting the file, but instead sending to stdout.
This should fix the issue in the Docker update_flow_provider script.

Note: This replicates a fix also made in pull requests #132 and #150 on github.

This closes #152.
This closes #150.
This closes #132.

Co-authored-by: Joao C Costa <joaocc@users.noreply.github.com>
Co-authored-by: silver2k <colin@findlay.geek.nz>
Signed-off-by: Kevin Doran <kdoran@apache.org>